+++ sator/adapters/driven/repositories/vulnerability/mappers.py
@@ -1,0 +1,35 @@
+from osvutils.types.osv import OSV
+from nvdutils.models.cve import CVE
+
+from sator.core.models.vulnerability import Vulnerability
+
+
+class NVDVulnerabilityMapper:
+    @staticmethod
+    def map_nvd_vulnerability(cve: CVE) -> Vulnerability:
+        return Vulnerability(
+            id=cve.id,
+            assigner=cve.source,
+            description=cve.descriptions.get_eng_description().value,
+            published_date=cve.published_date
+        )
+
+
+class OSVVulnerabilityMapper:
+    @staticmethod
+    def map_osv_vulnerability(osv: OSV) -> Vulnerability:
+        return Vulnerability(
+            id=osv.id,
+            assigner="opensource@google.com",
+            description=OSVVulnerabilityMapper.get_description(osv),
+            published_date=osv.published
+        )
+
+    @staticmethod
+    def get_description(osv: OSV) -> str:
+        if osv.details:
+            return osv.details
+        elif osv.summary:
+            return osv.summary
+        else:
+            return ""
