+++ sator/adapters/driven/repositories/vulnerability/nvd.py
@@ -3,21 +3,10 @@ from typing import Dict, Any, List
 
 from sator.core.models.vulnerability import Vulnerability
 from sator.core.ports.driven.vulnerability_repository import VulnerabilityRepositoryPort
+from sator.adapters.driven.repositories.vulnerability.mappers import NVDVulnerabilityMapper
 
 from nvdutils.loaders.json.default import JSONDefaultLoader
 from nvdutils.data.profiles.zero_click import ZeroClickProfile
-from nvdutils.models.cve import CVE
-
-
-class NVDVulnerabilityMapper:
-    @staticmethod
-    def map_nvd_vulnerability(cve: CVE) -> Vulnerability:
-        return Vulnerability(
-            id=cve.id,
-            assigner=cve.source,
-            description=cve.descriptions.get_eng_description().value,
-            published_date=cve.published_date
-        )
 
 
 class NVDVulnerabilityRepository(VulnerabilityRepositoryPort):
