+++ sator/adapters/driven/repositories/vulnerability/osv.py
@@ -2,32 +2,12 @@ from typing import Dict, Any, List
 
 from sator.core.models.vulnerability import Vulnerability
 from sator.core.ports.driven.vulnerability_repository import VulnerabilityRepositoryPort
+from sator.adapters.driven.repositories.vulnerability.mappers import OSVVulnerabilityMapper
 
-from osvutils.types.osv import OSV
 from osvutils.core.loader import OSVDataLoader
 from osvutils.core.filters.loader import LoaderFilters
 from osvutils.core.filters.database import DatabaseFilter
 from osvutils.core.filters.affected_packages import AffectedPackagesFilter
-
-
-class OSVVulnerabilityMapper:
-    @staticmethod
-    def map_osv_vulnerability(osv: OSV) -> Vulnerability:
-        return Vulnerability(
-            id=osv.id,
-            assigner="opensource@google.com",
-            description=OSVVulnerabilityMapper.get_description(osv),
-            published_date=osv.published
-        )
-
-    @staticmethod
-    def get_description(osv: OSV) -> str:
-        if osv.details:
-            return osv.details
-        elif osv.summary:
-            return osv.summary
-        else:
-            return ""
 
 
 class OSVVulnerabilityRepository(VulnerabilityRepositoryPort):
